[INFOI-18866] Keep the undeleted CompletedCheckpoint handle on zookeeper

+++ flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/ZooKeeperCheckpointIDCounter.java
@@ -119,6 +119,11 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 
 	@Override
 	public void shutdown(JobStatus jobStatus) throws Exception {
+		shutdown(jobStatus, true);
+	}
+
+	@Override
+	public void shutdown(JobStatus jobStatus, boolean cleanCounterPath) throws Exception {
 		synchronized (startStopLock) {
 			if (isStarted) {
 				LOG.info("Shutting down.");
@@ -126,7 +131,7 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 
 				client.getConnectionStateListenable().removeListener(connectionStateListener);
 
-				if (jobStatus.isGloballyTerminalState()) {
+				if (jobStatus.isGloballyTerminalState() && cleanCounterPath) {
 					LOG.info("Removing {} from ZooKeeper", counterPath);
 					client.delete().deletingChildrenIfNeeded().inBackground().forPath(counterPath);
 				}
